Split removeTags
This function is actually two functions in one.
It has two different signature (different parameters and return type).
The Scrutinizer static analyzer gets confused with the return type.

Ideally, the public interface would have been better with two function,
but thit is now too late to change as it would be a breaking change.

Splitting it allows to better deal with that.