Factor common code in Promise
Unfortunately, `create` is different because the first parameter is
the result instead of an error.
All the functions have error first, then result.
This was a bad design choice, this does not align with with the way promises
work.
Promise have the resolve (result) first, which allow to make the reject
(error) optional.
This is hard to change right now because this is already part of the
public API. We need to write adapters to normalize all of that.
I suggest to normalize the functions internally and adapt the public API
to make it backward compatible.