Add support for UFID (unique file identifier) frame
See https://id3.org/id3v2.4.0-frames (Section 4.1) for details.

This frame is very similar to the `PRIV` frame, however the
implementation has been changed slightly.
Since the `identifier` is expected to be binary data, unlike the `PRIV`
frame, the implementation does not accept a string. This is left to
user to encode a string.
The `PRIV` frame implementation causes two issues:
- the encoding used to encode the string (currently utf8) is not
  documented and the user does not know how to decode the string
  received as a Buffer when read
- the create and read API are asymmetric

Like other ID3 constraints, this implementation does not try to
verify or apply any constraints on the data, for example the fact
that owner identifiers should be unique or that the size of the
identifier binary data should not be more than 64 bytes.

Note in the test and the documentation the usage of more expressive
example data than in the other frames.