Improve index.js docs and types
I wanted to use ErrorCallback name instead of WriteCallback but
it was conflicting with a system type, I have tried to use a JSDoc
namespace but could not get it working, so changed the name for now
too avoid wasting to much time to get it working.

See also
https://stackoverflow.com/questions/45836847/how-to-get-vs-code-to-understand-jsdocs-typedef-across-multiple-files/55767692

I think at some point it would be beneficial to use Typescript
and compile the project to JS at export time.