Fix jsmediatags type definition
jsmediatags references `XMLHttpRequest` which is part of the `dom`
library but we don't use this library.

Rather than including a library not available in node, a better
solution is to skip the library check.

https://stackoverflow.com/questions/48704043/exclude-types-typings-in-installed-dependencies

The error for reference:
node_modules/@types/jsmediatags/types/index.d.ts:6:11 - error TS2304: Cannot find name 'XMLHttpRequest'.

6     xhr?: XMLHttpRequest | undefined;